Move Client class into a separate file
Closes #11
